Rewrite suggest-grid to handle multiple meetings
This provides a major rewrite of the suggest-grid script to also support
scheduling sessions in more than one meeting.

The scheduling for multiple meetings remains rather simple for now. By default,
the script will merely schedule sessions in one meeting as before. But it has
`times`, it will try to use them. If possible, that's all fine. If not, it will
try to schedule the session during the same number of slots, but at different
times.

The logic to parse the `times` section from the issue description is still
missing though.

Also, the code is mostly untested for the time being, so likely full of bugs.
